(WIP) Started refactor. Only Gemini refactored for now, other crawlers broken. Extracted GeminiExample as a separate class
